CTFs/Ethernaut/Fallout_DONE/pyproject.toml
[tool.poetry]
name = "fallout"
version = "0.1.0"
description = ""
authors = ["SAIKAT KARMAKAR"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
eth-ape = "^0.6.9"
ape-alchemy = "^0.6.1"
ape-ganache = "^0.6.6"
ape-solidity = "^0.6.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
